move util functions from views into a new module
